feat: Add movie router and service
This commit adds a new movie router and service to the BookMyShow_Server project. The movie router handles GET requests for retrieving all movies, while the movie service fetches movies from the database using the Movie model. This feature will allow users to view a list of available movies on the BookMyShow website.