feat: Remove unused HTML files and update npm dependencies
This commit removes the unused HTML files for the plays, sports, and activities components in the BookMyShowClient application. These files are no longer needed as the corresponding components have been deleted. Additionally, the package.json file in the BookMyShowClient project has been updated to include the latest versions of the ngx-timeago and ngx-toastr npm dependencies. This ensures that the application is using the latest stable versions of these libraries, which may include bug fixes and new features.