feat: Add admin guard for protected routes
This commit adds an admin guard to the BookMyShowClient application to protect certain routes that should only be accessible to administrators. The `AdminGuard` class has been created, which implements the `CanActivate` interface from the Angular router. It checks if the current user is an admin by calling the `getCurrentUserId()` method from the `AuthServiceService`. If the user is not an admin, they will be redirected to the login page. This feature enhances the security of the application by restricting access to sensitive admin-only functionality.